pyproject.toml
[tool.poetry]
name = "cat_task"
version = "0.1.0"
description = ""
authors = ["Sirojiddin Yakubov <yakubov9791999@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "4.8.0"
djoser = "^2.1.0"
django-cors-headers = "^3.12.0"
python-dotenv = "^0.20.0"
django-filter = "^21.1"
psycopg2-binary = "^2.9.3"
drf-yasg = "^1.20.0"
Pillow = "^9.1.1"
python-dateutil = "^2.8.2"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
